etl/experimental/bepinex-gen1/requirements.txt
# BepInEx Gen1 - Python Dependencies

# Core dependencies
pyyaml>=6.0
numpy>=1.24.0
pillow>=10.0.0
docker>=6.1.0

# Data processing
scipy>=1.11.0
pandas>=2.0.0

# Validation
jsonschema>=4.19.0

# Logging and monitoring
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
